+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 - 2019-09-29: Remove INDYGO and GlobalStaticOptimization from Moco.
 
+- 2019-09-23: MocoControlGoal has properties "exponent" and
+              "divide_by_displacement". 
+
+- 2019-09-05: MocoAverageSpeedGoal was added.
+
 - 2019-08-29: Users now interface with num_mesh_intervals instead of
               num_mesh_points when setting times to sample.